requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
matplotlib>=3.5.0
albumentations>=1.3.0
tqdm>=4.65.0
timm>=0.9.0
